add estimated_value_override to the program type
